LayoutTests/fast/forms/submit-button-submit-event-expected.txt

Tests that SubmitEvent.submitter property is properly set for .
On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS submitEvent.submitter is testButton
PASS successfullyParsed is true

TEST COMPLETE
